plugin

+++ ejercicio/_08-serial-port/_08-serial-port.ino
@@ -6,6 +6,6 @@ void setup() {
 }
 
 void loop() {
-  float value = analogRead(potenPin);
+  int value = analogRead(potenPin);
   Serial.println(value);
 }
